Change 32-bit to 64-bit system

+++ src/Basic/Type/data_type.hpp
@@ -2,12 +2,12 @@
 
 #include <stdint.h>
 
-typedef uint32_t count_type;
-typedef uint32_t vertex_id_type;
-typedef uint32_t edge_data_type;
+typedef uint64_t count_type;
+typedef uint64_t vertex_id_type;
+typedef uint64_t edge_data_type;
 typedef uint32_t degree_type;
-typedef uint32_t countl_type;      // uint64_t
-typedef uint32_t vertex_data_type; // double
+typedef uint64_t countl_type;
+typedef uint64_t vertex_data_type;
 
 typedef uint64_t address_type;
 typedef uint8_t machine_type;
